Hopefully the final v1.0 commit
Should be ready to be released.

+++ IVSDKDotNet/IVSDK/CVector.h
@@ -3,11 +3,11 @@ class CVector
 public:
 	float x, y, z;
 
-	float Heading() { return atan2(-x, y); }
+	float Heading() { return (float)atan2(-x, y); }
 	float MagnitudeSqr() { return x * x + y * y + z * z; }
 	float MagnitudeSqr2D() { return x * x + y * y; }
-	float Magnitude() { return sqrt(MagnitudeSqr()); }
-	float Magnitude2D() { return sqrt(MagnitudeSqr2D()); }
+	float Magnitude() { return (float)sqrt(MagnitudeSqr()); }
+	float Magnitude2D() { return (float)sqrt(MagnitudeSqr2D()); }
 };
 
 class CVector_pad : public CVector
